update big decimal

+++ app/src/main/java/com/hqnguyen/syl_v2/data/InfoTracking.kt
@@ -1,7 +1,9 @@
 package com.hqnguyen.syl_v2.data
 
+import java.math.BigDecimal
+
 data class InfoTracking(
     var speed: Float = 0f,
     var kCal: Float = 0f,
-    var distance: Float = 0f
+    var distance: BigDecimal =  BigDecimal(0)
 )
